requirements.txt
streamlit==1.32.0
python-docx==1.1.0
PyPDF2==3.0.1
requests==2.31.0 